scripts/requirements.txt
dotenv==0.9.9
langsmith==0.3.38
openai==1.76.0
langchain_core==0.3.56
python-slugify==8.0.4
